feat: refactor categories to match new beta changes
