+++ src/main/java/dev/dfonline/flint/actiondump/categories/ValueCategory.java
@@ -1,8 +1,8 @@
-package dev.dfonline.flint.actiondump.gamevalues;
+package dev.dfonline.flint.actiondump.categories;
 
 import dev.dfonline.flint.actiondump.data.Icon;
 
-public record GameValueCategory(
+public record ValueCategory(
         String identifier,
         int guiSlot,
         Icon icon
