--- src/main/java/dev/dfonline/flint/actiondump/sound/SoundCategory.java
@@ -1,10 +1,0 @@
-package dev.dfonline.flint.actiondump.sound;
-
-import dev.dfonline.flint.actiondump.data.Icon;
-
-public record SoundCategory(
-        String identifier,
-        Icon icon,
-        boolean hasSubCategories
-) {
-}
